fix: Add directory router
