+++ server/src/router/mod.rs
@@ -1,0 +1,2 @@
+pub mod teams;
+pub mod default;
